requirements/requirements.txt
ovos-plugin-manager>=0.0.26a5
python-vlc>=1.1.2